+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 * Add a heartbeat to reset a test's timeout whenever the test interacts with the
   test infrastructure.
 
+* `expect()`, `expectAsync()`, and `expectAsyncUntil()` throw more useful errors
+  if called outside a test body.
+
 ## 0.12.2
 
 * Convert JavaScript stack traces into Dart stack traces using source maps. This